+++ gulp/watch.js
@@ -1,0 +1,6 @@
+var gulp = require('gulp');
+
+// Watch tasks
+gulp.task('watch', function() {
+  gulp.watch('scss/**/*', ['sass:develop']);
+});
